move routeGenerator to api folder
