+++ backend/api/routeGenerator.js
@@ -1,3 +1,4 @@
+const { handleError } = require("./utilities");
 /*
 The API looks for actions specified by "funct" or "action" in
 either GET or POST requests.  You should pass along the appropriate
