+++ backend/api/user.js
@@ -1,5 +1,5 @@
 const passport = require("passport");
-const routeGenerator = require("../routeGenerator");
+const routeGenerator = require("./routeGenerator");
 const { handleError } = require("./utilities");
 const crypto = require("crypto");
 const userDB = require("../db/user");
